update to BETA-0.1.0, in this version the A9GTools scripts were improved, a local python interpreter was added and the compilation rules in the makefiles were modified to avoid dependencies on the original sdk locations.

+++ SDK/include/std_inc/stddef.h
@@ -14,11 +14,11 @@
 
 #ifndef _USERGEN
 #ifndef size_t
-typedef unsigned int                size_t;
+typedef unsigned long               size_t;
 #endif
 
 #ifndef ssize_t
-typedef int                         ssize_t;
+typedef long                        ssize_t;
 #endif
 #define SSIZE_MAX INT_MAX
 #endif
